// src/app/shared/models/e-commerce/company-info.model.ts
export class CompanyInfo{
    companyInfoId: number;
    companyName: string;
    phone1: number;
    phone2: number;
    email: string;
    instagramLink: string;
    facebookLink: string;
    linkedinLink: string;
    twitterLink: string;
    address: string;
    longitude: string;
    latitude: string;
}